Quit warning about one simulator when timeout occurs during launch.
For some reason, when starting a second simulator while another one
is already running, a modal dialog appears. The standard "quit" event
will not work.

If after a timeout occurs trying to contact the Frank server, let's
try clicking that button in that modal and see if next time we can
start the simulator normally.