Fix build failure due to treating frankified_build as the .app directory if it did not exist.
The copy command used to copy the built .app bundle into the frankified_build directory would only copy the .app into the frankified_build directory if the build directory already existed. If the build directory did not exist, the contents of the .app bundle would be copied into frankified_build.